fixed all logic so gameboard can be rendered after showing players their colors
